% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{make_lmer_formula}
\alias{make_lmer_formula}
\alias{make_tree_formula}
\title{Make New Formula}
\usage{
make_lmer_formula(response, fixed, random)

make_tree_formula(response, fixed)
}
\arguments{
\item{response}{Character string specifying the name of the response.}

\item{fixed}{Character string specifying the fixed effects structure.}

\item{random}{Character string specifying the random effects structure.}
}
\description{
Build model formulas character vectors describing the response name, fixed
effects structure, and (if required) random effects structure of the model.
}

