% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get_response_name}
\alias{get_fixed_formula}
\alias{get_random_formula}
\alias{get_response_name}
\title{Extract Formula Components}
\usage{
get_response_name(formula)

get_fixed_formula(formula)

get_random_formula(formula)
}
\arguments{
\item{formula}{A valid \code{mertree} model formula. See \code{\link{lmer}}
for details.}
}
\description{
Extract various components from a \code{mertree} model formula.
}

