% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nleaves}
\alias{nleaves}
\alias{nleaves.BinaryTree}
\alias{nleaves.rpart}
\title{Number of Leaves/Terminal Nodes}
\usage{
nleaves(object)

\method{nleaves}{rpart}(object)

\method{nleaves}{BinaryTree}(object)
}
\arguments{
\item{object}{An object of class \code{"rpart"} or \code{"BinaryTree"}.}
}
\value{
Returns \code{TRUE} if \code{object} does not contain any splits and
\code{FALSE} otherwise.
}
\description{
Compute the number of leaves (i.e., the number of temrinal nodes) of a tree.
}

