% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial_1d}
\alias{partial_1d}
\title{Partial Dependence (One Variable)}
\usage{
partial_1d(object, x.name, n, x.class, newdata, ...)
}
\arguments{
\item{object}{A \code{mertree} object.}

\item{x.name}{Character string giving the name of the independent variable of
interest.}

\item{n}{Integer giving the number of unique data points to use in
computing the partial dependence values.}

\item{x.class}{Character string specifying the class for \code{x.name}.}

\item{newdata}{An optional data frame.}

\item{...}{Additional optional arguments passed onto \code{aaply}.}
}
\description{
One- and two-dimensional partial dependence plots for objects of class
\code{mertree}.
}
\note{
It may be necessary to supply a value for \code{x.class} when \code{n} is
supplied. This is usually the case when \code{object} is of class
\code{"BinaryTree"} or \code{"mertree"} (with \code{unbiased = TRUE}).
}

