% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial_2d}
\alias{partial_2d}
\title{Partial Dependence (Two Variables)}
\usage{
partial_2d(object, x1.name, x2.name, n1, n2, x1.class, x2.class, newdata, ...)
}
\arguments{
\item{object}{A \code{mertree} object.}

\item{x1.name}{Character string giving the name of the first independent
variable of interest.}

\item{x2.name}{Character string giving the name of the second independent
variable of interest.}

\item{n1}{Integer giving the number of unique data points to use in
computing the partial dependence values for \code{x1.name}.}

\item{n2}{Integer giving the number of unique data points to use in
computing the partial dependence values for \code{x2.name}.}

\item{x1.class}{Character string specifying the class for \code{x1.name}.}

\item{x2.class}{Character string specifying the class for \code{x2.name}.}

\item{newdata}{An optional data frame.}

\item{...}{Additional optional arguments passed onto \code{aaply}.}
}
\description{
One- and two-dimensional partial dependence plots for objects of class
\code{mertree}.
}
\note{
It may be necessary to supply values for \code{x1.class} or \code{x2.class}
when \code{n1} or \code{n2}, respectively, are supplied. This is usually the
case when \code{object} is of class \code{"BinaryTree"} or \code{"mertree"}
(with \code{unbiased = TRUE}).
}

