% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simd}
\alias{simd}
\title{Simulated Data}
\format{A data frame with 500 rows and 9 variables}
\source{
Breiman, Leo (1996) Bagging predictors. Machine Learning 24, pages 123-140.

Friedman, Jerome H. (1991) Multivariate adaptive regression splines. The
Annals of Statistics 19 (1), pages 1-67.

Zhang, H.. (1997). Multivariate Adaptive Splines for Analysis of Longitudinal
Data. Journal of Computational and Graphical Statistics, 6(1), 74–91.
}
\description{
Data from Frieman (1991) and Breiman (1996). However, here the errors were
generated from a 5-dimensional Gaussian distribution with zero mean and a
variance-covariance matrix with 4 along the diagonal and 0.8 in the
off-diagonal; see Zhang (1997) for details. The function
\code{\link{mlbench.friedman1}} in the \code{mlbench} package simulates data
from the same function, but with independent errors.
}
\details{
\itemize{
  \item \code{subject}: Subject ID.
  \item \code{x1}: Baseline covariate.
  \item \code{x2}: Baseline covariate.
  \item \code{x3}: Baseline covariate.
  \item \code{x4}: Baseline covariate.
  \item \code{x5}: Baseline covariate.
  \item \code{x6}: Time-dependent covariate.
  \item \code{time}: Time variable.
  \item \code{y}: Response variable.
}
}
\keyword{datasets}

