% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_node.R
\name{assign_node}
\alias{assign_node}
\alias{assign_node.BinaryTree}
\alias{assign_node.mertree}
\alias{assign_node.rpart}
\title{Terminal Node Assignment}
\usage{
assign_node(object, newdata, na.action, ...)

\method{assign_node}{rpart}(object, newdata, na.action = na.rpart, ...)

\method{assign_node}{BinaryTree}(object, newdata, ...)

\method{assign_node}{mertree}(object, newdata, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{rpart}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{na.action}{The default action deletes all observations for which the
response is missing, but keeps those in which one or more predictors are
missing.}

\item{...}{Additional optional arguments. At present, no optional arguments
are used.}
}
\value{
A numeric vector containing the terminal node each observation
  belongs to.
}
\description{
Assign observations to a terminal node.
}

