% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mertree.R
\name{mertree}
\alias{mertree}
\title{Regression Trees for Longitudinal and Clustered Data}
\usage{
mertree(formula, data, unbiased = TRUE, initial_re, REML = TRUE,
  lmer.control = lmerControl(calc.derivs = FALSE), tree.control = if
  (unbiased) ctree_control() else rpart.control(), cv = TRUE, tol = 0.001,
  maxiter = 100L, do.trace = FALSE)
}
\arguments{
\item{formula}{An appropriate \code{\link{lmer}}-style formula.}

\item{data}{An optional data frame containing the variables named in
\code{formula}.}

\item{unbiased}{Logical indicating whether or not to use a conditional
inference tree. Default is \code{TRUE}.}

\item{initial_re}{Numeric vector containing the initial values for the random
effects. If omitted then defaults to zero.}

\item{REML}{Logical indicating whether or not the estimates should be chosen
to optimize the REML criterion (as opposed to the log-likelihood).}

\item{lmer.control}{List of control parameters for \code{\link{lmer}}.}

\item{tree.control}{List of control parameters for \code{\link{ctree}} or
\code{\link{rpart}}.}

\item{cv}{Logical indicating whether or not to prune each tree based on
cross-validations. Only used when \code{unbiased = FALSE}. Default is
\code{TRUE}.}

\item{tol}{The desired accuracy (convergence tolerance). Default is
\code{0.001)}}

\item{maxiter}{Integer specifying the maximum number of iterations. Default
is \code{1000}.}

\item{do.trace}{Logical indicating whether or not to print trace information.}
}
\description{
Fits an unbiased regression tree to longitudinal or clusteredl data by
iterating back and forth between a conditional inference regression tree to
capture complex interactions and nonlinear relationaships and a linear
mixed-effects model to capture complex correlation structure.
}

