% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial_2d}
\alias{partial_2d}
\title{Partial Dependence (Two Variables)}
\usage{
partial_2d(object, x1.name, x2.name, n1, n2, newdata)
}
\arguments{
\item{object}{A \code{mertree} object.}

\item{x1.name}{Character string giving the name of the first independent
variable of interest.}

\item{x2.name}{Character string giving the name of the second independent
variable of interest.}

\item{n1}{Integer giving the number of unique data points to use in
computing the partial dependence values for \code{x1.name}.}

\item{n2}{Integer giving the number of unique data points to use in
computing the partial dependence values for \code{x2.name}.}

\item{newdata}{An optional data frame.}
}
\description{
One- and two-dimensional partial dependence plots for objects of class
\code{mertree}.
}

