% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simd}
\alias{simd}
\title{Simulated Data}
\format{A data frame with 500 rows and 9 variables}
\source{
[REF1]
[REF2]
}
\description{
The data are described in [REF1] and (originally) in[REF2].
}
\details{
\itemize{
  \item \code{subject} Subject ID.
  \item \code{x1} First predictor variable.
  \item \code{x2} Second predictor variable.
  \item \code{x3} Third predictor variable.
  \item \code{x4} Fourth predictor variable.
  \item \code{x5} Fifth predictor variable.
  \item \code{x6} Sixth predictor variable.
  \item \code{time} Time variable.
  \item \code{y} Response variable.
}
}
\keyword{datasets}

