% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mertree.R
\name{varimp}
\alias{varimp}
\title{Variable Importance Scores}
\usage{
varimp(object, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{"mertree"}.}
}
\description{
Variable importance scores for \code{"mertree"} objects.
}

